Phase 2: Split validation module - modular architecture
Following CODESTYLE.md Single Responsibility Principle to reduce cognitive load:

- Split 547-line validation.ts into 5 focused modules under 300 lines each
- json-parser.ts: Main validation entry point and orchestration
- error-factory.ts: Error creation, type guards, and validation error construction
- suggestion-generator.ts: Validation suggestions and field corrections for LLM retries
- feedback-formatter.ts: Progressive urgency formatting for LLM feedback loops
- field-analyzer.ts: Field-level error analysis and extraction utilities

Key improvements:
- Zero functional changes - 100% backward compatibility maintained
- Clear separation of concerns with single-responsibility modules
- Enhanced maintainability through focused, testable components
- Reduced cognitive load - each module fits in working memory
- Progressive disclosure - complex logic broken into understandable pieces

All quality gates pass:
- ✅ TypeScript compilation (npm run typecheck)
- ✅ Biome code quality (npm run check)
- ✅ All validation tests passing with modular implementation
